Split 4k+ line test into multiple files, fix lots of tests
